3a

+++ src/xrGame/script_game_object3.cpp
@@ -1469,7 +1469,7 @@ u8 CScriptGameObject::GetRestrictionType()
 	if (restr)
 		return restr->m_space_restrictor_type;
 
-	return (-1);
+	return u8(-1);
 }
 
 void CScriptGameObject::SetRestrictionType(u8 typ)
